+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Show ToeTacTic ASCII art banner when the game is loaded
 
+### Changed
+- Limit board dimension to between 3 and 8
+
 ### Fixed
 - Remove duplicate opponent list from the choose opponent prompt